Flip default setting to "true" for three flags:
- apm_config.peer_service_aggregation
- apm_config.peer_tags_aggregation
- apm_config.compute_stats_by_span_kind

This is to implement the GA rollout of Service Representation, which has been in Limited Preview until now.

This is a non-breaking change: all back-end data and pipelines are already updated, and inferred entities are appearing throughout APM. This setting ensures that those inferred entities will have stats properly aggregated and attributed to them.